Update 68CLISP.ASM
Found errors in the Error message section of the code where the original fcs directive is not supported in the Motorola assembler.  It looks like that was relic of an assembler for the flex OS, which to save string space, would set the sign bit on the last character to indicate the end of the string.  I'll have to try this out at a later time, but here is a possible coding that may work with the FCB directive and ORing the character with a $80H to set the sign bit.  Don't know if this is legal syntax or not yet.  Other possibility would be to just use the hex ascii code for the character and add the high bit with the character.  That's not nearly as readable, however.  i.e. R would be 0101 0010 OR 1000 0000 => FCB %11010010. Ugly.